requirements.txt
pillow
pillow-heif
piexif
customtkinter
pyinstaller